% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNet.R
\name{autocurve.edges2}
\alias{autocurve.edges2}
\title{Adjust Edge Curvature for Visualization}
\usage{
autocurve.edges2(graph, start = 0.5)
}
\arguments{
\item{graph}{An \code{igraph} object.}

\item{start}{Numeric value indicating the starting curvature.}
}
\value{
A numeric vector of curvature values for edges.
}
\description{
This function calculates the curvature of edges in a graph for better visualization.
}
