% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeWeight.R
\name{Calculate_sum_of_weights}
\alias{Calculate_sum_of_weights}
\title{Calculate Sum of Weights for Network Nodes}
\usage{
Calculate_sum_of_weights(network, offset = FALSE, log = FALSE, abs = TRUE)
}
\arguments{
\item{network}{Data frame. Input network with columns for nodes and weight values.}

\item{offset}{Logical. If TRUE, removes the first column if it’s an index.}

\item{log}{Logical. If TRUE, applies a log transformation to the result.}

\item{abs}{Logical. If TRUE, calculates the sum of absolute weights.}
}
\value{
Data frame. Sum of weights for each node across samples.
}
\description{
Calculates the sum of edge weights (absolute or non-absolute) for each node in a network across multiple samples.
}
