% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSN_Properties.R
\name{determineCharacteristics}
\alias{determineCharacteristics}
\title{Determine Network Characteristics}
\usage{
determineCharacteristics(network)
}
\arguments{
\item{network}{Data frame representing the network, where the first two columns are edges and subsequent columns are sample weights.}
}
\value{
Saves local and global network properties as .tsv files and returns a data frame of global properties.
}
\description{
This function calculates local and global properties for each sample in a network.
It extracts subgraphs with non-zero weights, computes node-level properties, and saves results.
}
