% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{filter_OTU}
\alias{filter_OTU}
\title{过滤、筛选OTU
@param otu_table 过滤后的otu表
@param Top 过滤原始表格，top>=1时按照top值选取平均丰度前top，1>top>0时，选取平均丰度高于top值
@param Pre 物种流行度阈值
@export}
\usage{
filter_OTU(otu_table, Top = NULL)
}
\description{
过滤、筛选OTU
@param otu_table 过滤后的otu表
@param Top 过滤原始表格，top>=1时按照top值选取平均丰度前top，1>top>0时，选取平均丰度高于top值
@param Pre 物种流行度阈值
@export
}
