% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterOTU2.R
\name{filter_OTU2}
\alias{filter_OTU2}
\title{Filter OTUs Based on Prevalence}
\usage{
filter_OTU2(otu_table, Pre = NULL)
}
\arguments{
\item{otu_table}{Data frame where rows are OTUs and columns are samples.}

\item{Pre}{Numeric value between 0 and 1. Filters OTUs that are present in at least \verb{Pre * 100\%} of samples.}
}
\value{
A filtered OTU table.
}
\description{
This function filters OTUs based on their prevalence across samples.
}
