% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrobalProCompare.R
\name{generate_random_network}
\alias{generate_random_network}
\title{生成随机网络}
\usage{
generate_random_network(graph, type = "gnm")
}
\arguments{
\item{graph}{\code{igraph}对象，输入的网络}

\item{type}{字符串，随机网络的类型}
}
\value{
\code{igraph}对象，生成的随机网络
}
\description{
生成随机网络
}
