% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransEdge.R
\name{get_full_edge_list}
\alias{get_full_edge_list}
\title{Get Full Edge List from Matrix}
\usage{
get_full_edge_list(mat)
}
\arguments{
\item{mat}{Numeric matrix.}
}
\value{
A data frame containing edge pairs and weights.
}
\description{
This function converts a matrix into an edge list data frame.
}
