% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BridgeNet.R
\name{get_top_bridge_nodes}
\alias{get_top_bridge_nodes}
\title{Get Top Bridge Nodes}
\usage{
get_top_bridge_nodes(bridge_centrality)
}
\arguments{
\item{bridge_centrality}{List, bridge centrality analysis result containing bridge strength values.}
}
\description{
Extracts nodes with top bridge strength.
}
