% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrobalPropCompare.R
\name{grobal_pro_compare}
\alias{grobal_pro_compare}
\title{与随机生成的网络进行全局属性比较}
\usage{
grobal_pro_compare(
  graph = graph,
  type = "gnm",
  step = 100,
  netName = "CK",
  ncpus = ncpus
)
}
\description{
@param graph 输入需要比较的网络
@param type 随机网络的类型
@param step 迭代次数
@param netName 随机网络的名字
@param ncpus 运行设置的线程
@export
}
