% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeTable.R
\name{merge_bio}
\alias{merge_bio}
\title{Merge and Filter OTU Tables from Two Domains}
\usage{
merge_bio(
  table1,
  table2,
  table1name = "domain1",
  table2name = "domain2",
  top = NULL,
  pvl.threshold = 0.5
)
}
\arguments{
\item{table1}{Data frame of the first OTU table.}

\item{table2}{Data frame of the second OTU table.}

\item{table1name}{Character string specifying the prefix name for species in \code{table1}. Default is "domain1".}

\item{table2name}{Character string specifying the prefix name for species in \code{table2}. Default is "domain2".}

\item{top}{Numeric value. If \code{top >= 1}, selects the top \code{top} species based on average abundance.
If \verb{0 < top < 1}, selects species with average abundance greater than \code{top}.}

\item{pvl.threshold}{Numeric value between 0 and 1 for prevalence threshold in filtering species.}
}
\value{
Data frame of merged OTU tables with filtered data.
}
\description{
This function processes and merges two OTU tables with filtering options.
}
