% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modularity.R
\name{modularity_igraph}
\alias{modularity_igraph}
\title{Network Modularity Calculation}
\usage{
modularity_igraph(g, clu_method = "cluster_fast_greedy")
}
\arguments{
\item{g}{An \code{igraph} object representing the network.}

\item{clu_method}{Character string specifying the network clustering method. Supported methods are "cluster_walktrap", "cluster_edge_betweenness", "cluster_fast_greedy", "cluster_spinglass".}
}
\value{
A list containing the membership vector and the modularity value.
}
\description{
This function calculates the modularity of a network and returns the membership of nodes and the modularity value.
}
