% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline2.R
\name{network_pipeline}
\alias{network_pipeline}
\title{基于微生物相关性构建微生物网络以及后续相关分析}
\usage{
network_pipeline(
  table1,
  tax = NULL,
  cor_table_list = NULL,
  group_df = NULL,
  pvl.threshold = 0,
  r.threshold = 0.6,
  p.threshold = 0.05,
  method = "spearman",
  R = 10,
  ncpus = 10,
  clu_method = "cluster_fast_greedy",
  p.adj = "fdr",
  step = 100,
  top = NULL,
  rm.p.list = seq(0.05, 0.95, by = 0.05),
  zipi = FALSE,
  compare_net = FALSE,
  features = FALSE
)
}
\description{
@param table1 研究对象的丰度矩阵（标准化后）
@param tax 分类注释信息
@param cor_table_list 待分析的相关性矩阵列表
@param group_df 研究对象分组信息表，NULL自动按照样品名生成，FALSE默认为只有一组样品
@param pvl.threshold 物种流行度阈值
@param r.threshold 相关性R值阈值
@param p.threshold 相关性P值阈值
@param method 相关性网络构建方法，spearman、pearson、sparcc、spieceasi四种方法
@param R SSPCC构建网络迭代参数
@param ncpus 运行设置的线程
@param clu_method 网络聚类方法
@param p.adj 网络相关性的多重矫正方法
@param step 迭代次数
@param top 过滤原始表格，top>=1时按照top值选取平均丰度前top，1>top>0时，选取平均丰度高于top值
@param rm.p.list 模拟随机毁灭的比例
@param zipi 对节点进行zipi值计算
@param node.cluster 每个模块点数量阈值，大于这个阈值才上色
@param conpare_net 与随机空白网络进行比较
@param features 是否进行鲁棒性、脆弱性模拟
@examples
network_pipeline(otu,group_df = FALSE,zipi = TRUE,compare_net = TRUE,features = TRUE)
@return 输出网络矩阵以及一系列分析结果和可视化
@export
}
