% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Properties.R
\name{node_properties}
\alias{node_properties}
\title{Calculate Network Properties}
\usage{
node_properties(
  graph,
  clu_method = "cluster_walktrap",
  zipi = FALSE,
  RM = TRUE,
  tag = "NA",
  output = "./zipi/"
)
}
\arguments{
\item{graph}{An \code{igraph} object representing the network.}

\item{clu_method}{Character, network clustering method. Default is "cluster_walktrap".}

\item{zipi}{Logical, whether to calculate ZiPi values. Default is FALSE.}

\item{RM}{Logical, whether to calculate relative modularity. Default is TRUE.}

\item{tag}{Character, network name. Default is "NA".}

\item{output}{Character, output file path. Default is "./zipi/".}
}
\value{
A list containing local and global indicators.
}
\description{
This function calculates local and global properties of a given network.
}
