% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Properties2.R
\name{node_properties2}
\alias{node_properties2}
\title{网络属性计算}
\usage{
node_properties2(
  igraph,
  clu_method = "cluster_walktrap",
  zipi = FALSE,
  RM = TRUE,
  tag = "NA",
  output = "./zipi/"
)
}
\arguments{
\item{igraph}{\code{igraph}对象，待分析的网络图}

\item{clu_method}{字符串，网络聚类方法，默认为 "cluster_walktrap"}

\item{zipi}{逻辑值，是否计算Zipi值，默认为TRUE}

\item{RM}{逻辑值，是否计算相对模块化，默认为TRUE}

\item{tag}{字符串，网络名称，默认为NULL}

\item{output}{字符串，输出文件路径，默认为 "./zipi/"}
}
\value{
返回包含局部和全局指标的列表
}
\description{
网络属性计算
}
