% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPCOA.R
\name{PCA_draw}
\alias{PCA_draw}
\title{PCA and PCoA Plotting Function}
\usage{
PCA_draw(
  table,
  group_df,
  vscol,
  save = TRUE,
  showType = NULL,
  mode = "PCOA",
  offset = FALSE
)
}
\arguments{
\item{table}{Data frame. The input data where columns represent samples.}

\item{group_df}{Data frame. Group metadata with sample names and a grouping column.}

\item{vscol}{Character. Column name in \code{group_df} indicating the group variable.}

\item{save}{Logical. If TRUE, saves the plot as PDF.}

\item{showType}{Character. Options for additional plot features: \code{"centroid"} or \code{"ellipse"}. Default is NULL.}

\item{mode}{Character. Analysis mode: \code{"PCA"} or \code{"PCOA"}. Default is \code{"PCOA"}.}

\item{offset}{Logical. If TRUE, combines two columns into "edges" in the input table. Default is FALSE.}
}
\description{
This function generates 2D PCA or PCoA plots for structured data, allowing for centroid and ellipse visualizations.
}
