% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BridgeNet.R
\name{perform_prediction}
\alias{perform_prediction}
\title{Perform Prediction Analysis}
\usage{
perform_prediction(data_matrix)
}
\arguments{
\item{data_matrix}{Matrix, the numeric data matrix used for prediction.}
}
\value{
A vector of prediction error values.
}
\description{
Uses mgm to perform predictive analysis on a matrix.
}
