% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BridgeNet.R
\name{plot_bridge_centrality}
\alias{plot_bridge_centrality}
\title{Plot Bridge Centrality}
\usage{
plot_bridge_centrality(bridge_centrality)
}
\arguments{
\item{bridge_centrality}{List, bridge centrality analysis results.}
}
\description{
Visualizes bridge centrality using specified metrics.
}
