% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BridgeNet.R
\name{plot_bridge_network}
\alias{plot_bridge_network}
\title{Bridge Network Plot}
\usage{
plot_bridge_network(network, groups, colors, prediction)
}
\arguments{
\item{network}{Matrix, sub-network for bridge analysis.}

\item{groups}{List, positions of nodes in each group.}

\item{colors}{Named vector of colors for each group.}

\item{prediction}{Vector, prediction error values for pie charts.}
}
\description{
Bridge Network Plot
}
