% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BridgeNet.R
\name{plot_main_network}
\alias{plot_main_network}
\title{Main Network Plot}
\usage{
plot_main_network(network, groups, colors)
}
\arguments{
\item{network}{Matrix, filtered correlation matrix.}

\item{groups}{List, positions of groups in the network.}

\item{colors}{Named vector of colors for each group.}
}
\description{
Main Network Plot
}
