% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRobVul.R
\name{plot_robustness_summary}
\alias{plot_robustness_summary}
\title{Plot Robustness Summary}
\usage{
plot_robustness_summary(rrob_sum)
}
\arguments{
\item{rrob_sum}{Data frame containing robustness summary data.}
}
\value{
A ggplot object.
}
\description{
Plot Robustness Summary
}
