% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRobVul.R
\name{plot_vulnerability_detail}
\alias{plot_vulnerability_detail}
\title{Plot Vulnerability Detail}
\usage{
plot_vulnerability_detail(rrob_detail, compare_groups = NULL)
}
\arguments{
\item{rrob_detail}{Data frame containing the detailed vulnerability data.}

\item{compare_groups}{List of group comparisons for statistical testing (optional).}
}
\value{
A \code{ggplot} object representing the vulnerability plot.
}
\description{
This function generates a detailed plot of network vulnerability under random node removal.
}
