% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNet.R
\name{plotnetwork}
\alias{plotnetwork}
\title{Network Visualization}
\usage{
plotnetwork(
  g,
  clu_method = "cluster_fast_greedy",
  tag = "network",
  df = NULL,
  node_cluster = 10,
  output_dir = "./"
)
}
\arguments{
\item{g}{An \code{igraph} object representing the network.}

\item{clu_method}{Character string specifying the network clustering method.}

\item{tag}{Character string specifying the network name.}

\item{df}{Optional data frame containing species abundance data to size nodes.}

\item{node_cluster}{Integer, the minimum number of nodes in a module to assign a unique color.}

\item{output_dir}{Character string specifying the output directory.}
}
\description{
This function visualizes the network with nodes colored by modules and sized by average abundance.
}
