% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxPlot.R
\name{plotnetwork_tax}
\alias{plotnetwork_tax}
\title{网络可视化，根据物种信息上色
@param g 输入需要可视化的网络
@param tag 网络名
@param tax 物种分类信息
@param df 微生物丰度文件，提供则按照丰度中位数作为网络点的大小
@export}
\usage{
plotnetwork_tax(g, tag = tag, tax = tax, df = sel_species_df1)
}
\description{
网络可视化，根据物种信息上色
@param g 输入需要可视化的网络
@param tag 网络名
@param tax 物种分类信息
@param df 微生物丰度文件，提供则按照丰度中位数作为网络点的大小
@export
}
