% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotVolcano.R
\name{PlotVolcano}
\alias{PlotVolcano}
\title{Volcano Plot Function}
\usage{
PlotVolcano(total_res, typename)
}
\arguments{
\item{total_res}{Data frame containing differential analysis results.}

\item{typename}{Character string indicating the type of data ("Nodes" or "Edges").}
}
\value{
A ggplot2 object of the volcano plot.
}
\description{
Generate a volcano plot for visualizing differential expression across groups.
}
