% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatMatrix.R
\name{reorder_cor_and_p}
\alias{reorder_cor_and_p}
\title{Reorder Correlation and P-value Matrices}
\usage{
reorder_cor_and_p(cormat, pmat)
}
\arguments{
\item{cormat}{Correlation matrix.}

\item{pmat}{P-value matrix.}
}
\value{
A list containing reordered correlation and p-value matrices.
}
\description{
This function reorders the correlation and p-value matrices based on hierarchical clustering.
}
