% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rob.R
\name{Robustness.Random.removal}
\alias{Robustness.Random.removal}
\title{网络模块化
@param cor 输入网络相关性矩阵
@param otu 输入物种丰度表
@param tag 网络名字
@param nperm 迭代次数
@param rm.p.list 模拟随机毁灭的比例
@param ncpus 运行设置的线程
@param loc_pro 网络属性
@param zipi 是否根据zipi值作为target
@param bet_centrality 是否根据bet_centralityi值作为target
@param clo_centrality 是否根据clo_centrality值作为target
@param eig_centrality 是否根据eig_centrality值作为target
@param pagerank 是否根据pagerank值作为target
@param kcore 是否根据kcore值作为target
@param degree 是否根据degree值作为target
@export}
\usage{
Robustness.Random.removal(
  cor,
  otu,
  tag = tag,
  nperm = 100,
  rm.p.list = seq(0.05, 1, by = 0.05),
  ncpus = ncpus
)
}
\description{
网络模块化
@param cor 输入网络相关性矩阵
@param otu 输入物种丰度表
@param tag 网络名字
@param nperm 迭代次数
@param rm.p.list 模拟随机毁灭的比例
@param ncpus 运行设置的线程
@param loc_pro 网络属性
@param zipi 是否根据zipi值作为target
@param bet_centrality 是否根据bet_centralityi值作为target
@param clo_centrality 是否根据clo_centrality值作为target
@param eig_centrality 是否根据eig_centrality值作为target
@param pagerank 是否根据pagerank值作为target
@param kcore 是否根据kcore值作为target
@param degree 是否根据degree值作为target
@export
}
