% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaleSSNetwork.R
\name{scale_network}
\alias{scale_network}
\title{Scale Network Weights}
\usage{
scale_network(network, offset = FALSE, vose = FALSE, group = "Total")
}
\arguments{
\item{network}{Data frame. The input network with columns for edges and weight values.}

\item{offset}{Logical. If TRUE, removes the first column (for cases with unnecessary index column).}

\item{vose}{Logical. If TRUE, saves the scaled network to a specified output file.}

\item{group}{Character. Used as the identifier in the output file name when \code{vose = TRUE}.}
}
\value{
Data frame. Scaled network with weights normalized to the range.
}
\description{
This function scales the weights of a network (correlation matrix) to be within the range .
It optionally saves the scaled network to a file.
}
