% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectTop_n_Edges.R
\name{select_top_edges}
\alias{select_top_edges}
\title{Select Top Edges from Network}
\usage{
select_top_edges(
  n,
  table,
  group = "Total",
  remove_zero_rows = FALSE,
  offset = FALSE,
  vose = TRUE
)
}
\arguments{
\item{n}{Integer. Number of top edges to select.}

\item{table}{Data frame. Input network with columns for edges and weight values.}

\item{group}{Character. Identifier used in the output file name when \code{vose = TRUE}.}

\item{remove_zero_rows}{Logical. If TRUE, removes rows where all edge weights are zero across samples.}

\item{offset}{Logical. If TRUE, removes the first column if it's an index.}

\item{vose}{Logical. If TRUE, saves the filtered network to a file.}
}
\value{
Data frame. Network with only the top \code{n} edges retained for each sample.
}
\description{
Selects the top \code{n} edges from a network based on edge weights for each sample.
Edges not in the top \code{n} are set to zero.
}
