% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssPCC.R
\name{sspcc_cal2}
\alias{sspcc_cal2}
\title{Comparison-based Single-Sample Network Construction}
\usage{
sspcc_cal2(sel_otu_table, group_df, ck = "CK", group = "group")
}
\arguments{
\item{sel_otu_table}{Data frame. The abundance table of selected OTUs.}

\item{group_df}{Data frame. Contains group and sample information.}

\item{ck}{Character. The name of the control group in \code{group_df}.}

\item{group}{Character. The column name in \code{group_df} specifying the group variable.}
}
\value{
None. Saves each sample's single-sample PCC to individual TSV files in the \verb{./ssn/sspcc/} directory.
}
\description{
Constructs single-sample networks by calculating PCC differences between each sample in the specified group
and a control group.
}
