% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vulnerability.R
\name{Vulnerability}
\alias{Vulnerability}
\title{Calculate Network Vulnerability}
\usage{
Vulnerability(cor)
}
\arguments{
\item{cor}{Numeric matrix representing the undirected correlation network.}
}
\value{
A numeric value representing the maximum node vulnerability in the network.
}
\description{
This function calculates the vulnerability of each node and returns the maximum vulnerability of the network.
}
