% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZiPiplot.R
\name{ZiPiplot}
\alias{ZiPiplot}
\title{Zi-Pi Plot Visualization}
\usage{
ZiPiplot(
  igraph,
  clu_method = "cluster_fast_greedy",
  output_dir = "./",
  tag = "group"
)
}
\arguments{
\item{igraph}{An \code{igraph} object representing the network.}

\item{clu_method}{Character string specifying the network clustering method. Supported methods are "cluster_walktrap", "cluster_edge_betweenness", "cluster_fast_greedy", "cluster_spinglass".}

\item{output_dir}{Character string specifying the output directory. Defaults to current working directory.}

\item{tag}{Character string specifying the network name. Defaults to "group".}
}
\value{
A data frame containing Zi and Pi values and node types.
}
\description{
This function visualizes the Zi-Pi distribution of nodes in a network and classifies them into roles.
}
