% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRobVul.R
\name{compare_rob_vul}
\alias{compare_rob_vul}
\title{鲁棒值脆弱值的可视化}
\usage{
compare_rob_vul(rrob_sum_list, rrob_detail_list)
}
\description{
@param rrob_sum_list 生成每组的鲁棒值的平均值
@param rrob_detail_list 生成每组的鲁棒值的详细值
@export
}
