% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatMarix.R
\name{corMicro}
\alias{corMicro}
\title{基于微生物相关性构建微生物网络}
\usage{
corMicro(
  table1,
  method = "spearman",
  p.adj = "BH",
  sel_group = "total",
  R = 10,
  ncpus = 10,
  r.threshold = 0.3,
  p.threshold = 0.05,
  vose = TRUE
)
}
\description{
@param table1 过滤后的otu表
@param method 相关性网络构建方法，spearman、pearson、sparcc、spieceasi四种方法
@param p.adj 网络相关性的多重矫正方法
@param sel_group 所需要分析的组
@param R SSPCC构建网络迭代参数
@param ncpus 运行设置的线程
@param r.threshold 相关性R值阈值
@param p.threshold 相关性P值阈值
@param vose 是否输出相关文件
@export
}
