% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterOTU1.R
\name{filter_OTU}
\alias{filter_OTU}
\title{Filter OTUs Based on Average Abundance}
\usage{
filter_OTU(otu_table, Top = NULL)
}
\arguments{
\item{otu_table}{Data frame where rows are OTUs and columns are samples.}

\item{Top}{Numeric value. If \code{Top >= 1}, selects the top \code{Top} OTUs based on average abundance.
If \verb{0 < Top < 1}, selects OTUs with average abundance greater than \code{Top}.}
}
\value{
A filtered OTU table.
}
\description{
This function filters OTUs (Operational Taxonomic Units) from the abundance table based on average abundance.
}
