% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrobalProCompare.R
\name{grobal_pro_compare}
\alias{grobal_pro_compare}
\title{与随机生成的网络进行全局属性比较}
\usage{
grobal_pro_compare(
  graph,
  type = "gnm",
  step = 100,
  netName = sel_group,
  ncpus = 1
)
}
\arguments{
\item{graph}{\code{igraph}对象，待比较的网络}

\item{type}{字符串，随机网络的类型，默认为 "gnm"}

\item{step}{整数，随机网络的迭代次数，默认为 100}

\item{netName}{字符串，网络名称，默认为 "CK"}

\item{ncpus}{整数，运行并行计算时的 CPU 数量}
}
\value{
列表，包含实际网络与随机网络的全局属性对比结果和拟合模型的 p 值
}
\description{
与随机生成的网络进行全局属性比较
}
