% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modularity.R
\name{modularity_igraph}
\alias{modularity_igraph}
\title{网络模块化
@param g 输入需要模块化的网络
@param clu_method 网络聚类方法
@export}
\usage{
modularity_igraph(g, clu_method = "cluster_fast_greedy")
}
\description{
网络模块化
@param g 输入需要模块化的网络
@param clu_method 网络聚类方法
@export
}
