% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Properties.R
\name{node_properties}
\alias{node_properties}
\title{网络属性计算
@param igraph 输入需要属性分析的网络
@param clu_method 网络聚类方法
@param zipi 是否计算zipi值
@param RM 是否计算相对模块化
@param tag 网络名字
@param output 输出路径
@export}
\usage{
node_properties(
  igraph,
  clu_method = "cluster_walktrap",
  zipi = FALSE,
  RM = TRUE,
  tag = NULL,
  output = "./zipi/"
)
}
\description{
网络属性计算
@param igraph 输入需要属性分析的网络
@param clu_method 网络聚类方法
@param zipi 是否计算zipi值
@param RM 是否计算相对模块化
@param tag 网络名字
@param output 输出路径
@export
}
