% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNet.R
\name{plotnetwork}
\alias{plotnetwork}
\title{网络可视化
@param g 输入需要可视化的网络
@param clu_method 网络聚类方法.
@param tag 网络名
@param df 微生物丰度文件，提供则按照平均丰度作为网络点的大小
@param node.cluster 每个模块点数量阈值，大于这个阈值才上色
@export}
\usage{
plotnetwork(
  g,
  clu_method = "cluster_fast_greedy",
  tag = tag,
  df = NULL,
  node.cluster = 10
)
}
\description{
网络可视化
@param g 输入需要可视化的网络
@param clu_method 网络聚类方法.
@param tag 网络名
@param df 微生物丰度文件，提供则按照平均丰度作为网络点的大小
@param node.cluster 每个模块点数量阈值，大于这个阈值才上色
@export
}
