% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZipiPlot.R
\name{ZiPiplot}
\alias{ZiPiplot}
\title{Zipi值分布可视化
@param igraph 输入需要可视化的网络
@param clu_method 网络聚类方法
@param output 输出路径
@param tag 网络名字
@export}
\usage{
ZiPiplot(
  igraph = igraph,
  clu_method = "cluster_fast_greedy",
  output = "./",
  tag = NULL
)
}
\description{
Zipi值分布可视化
@param igraph 输入需要可视化的网络
@param clu_method 网络聚类方法
@param output 输出路径
@param tag 网络名字
@export
}
